feat: create route to rent a product

+++ src/modules/products/products.service.ts
@@ -7,11 +7,16 @@ import { User } from '@modules/users'
 
 import { NotFoundError } from '../common/utils/errors'
 
+import { ValidationError } from '@modules/common/utils'
+import { RentProductPayload } from './payloads'
+import { UsersProductsService } from '../usersProducts/usersProducts.service'
+
 @Injectable()
 export class ProductsService {
   constructor(
     @InjectRepository(Product)
-    private readonly productsRepository: Repository<Product>
+    private readonly productsRepository: Repository<Product>,
+    protected readonly usersProductsService: UsersProductsService
   ) {}
 
   async getAvailableToRent(owner: User) {
@@ -28,27 +33,51 @@ export class ProductsService {
     return await this.productsRepository.save({ ...payload, owner: user })
   }
 
-  async update(id: number, payload: ProductFillableFields) {
-    try {
-      await this.productsRepository.findOneOrFail(id)
-    } catch {
-      throw new NotFoundError({ entity: 'Product' })
-    }
+  async update(id: number, payload: ProductFillableFields, user: User) {
+    await this.get({ id, owner: user })
 
     await this.productsRepository.update(id, payload)
   }
 
-  async destroy(id: number) {
-    try {
-      await this.productsRepository.findOneOrFail(id)
-    } catch {
-      throw new NotFoundError({ entity: 'Product' })
+  async destroy(id: number, user: User) {
+    const product = await this.get({ id, owner: user })
+
+    if (product.isLent) {
+      throw new ValidationError({
+        message: 'You cant delete a product that is lent'
+      })
     }
 
     await this.productsRepository.delete(id)
   }
 
-  async get(id: number) {
-    return await this.productsRepository.findOne({ id })
+  async get(query: Record<string, unknown>) {
+    try {
+      return await this.productsRepository.findOneOrFail(query)
+    } catch (error) {
+      throw new NotFoundError({ entity: 'Product' })
+    }
+  }
+
+  async rentProduct({ productId }: RentProductPayload, user: User) {
+    const desiredProduct = await this.get({ id: productId })
+
+    if (desiredProduct.isLent) {
+      throw new ValidationError({
+        message: 'Produto já foi emprestado para outra pessoa'
+      })
+    }
+
+    const productOwner = await desiredProduct.owner
+
+    if (productOwner.id === user.id) {
+      throw new ValidationError({
+        message: 'Você não pode alugar o próprio produto'
+      })
+    }
+
+    await this.usersProductsService.create({ product: desiredProduct, user })
+
+    await this.productsRepository.update(productId, { isLent: true })
   }
 }
